meson: vdec: Rewrite design
Try to modularize by splitting the vdec into its different VDECS,
as well as differentiating the codecs.

Each OUTPUT format must now declare ops for its associated vdec & codec.

It's still not as good as I'd like it, but some design details
remain to be uncovered when implementing future codecs.