meson: vdec: h264: Fix SEI data handling
Bitstreams that feature SEI data will trigger the uCode
to write it to yet another buffer.

Add support by handling it in the ISR, and allocating a buffer
so that the SEI doesn't get written to random memory.

This doesn't seem necessary on older SoC/uCode (m8b for instance..)