meson: vdec: HEVC + General updates
* HEVC:
 - Fix 2 bugs that caused corruption on select samples.
The first one is that we used the session width/height rather
than what the hardware tells us. But the hardware gives us aligned
width/height which can be different.
It's planned to check what the hardware tells us against what
userspace told us.

The second one is that we didn't output correctly the buffers
when multiple IDR frames were back to back, which then lead to stalling
or artifacts. The fix is in codec_hevc_update_referenced.

There remains one bug to be fixed that affects frame order in certain cases..

 - Start to add code to handle AML frame compression, which would also allow 10-bit support.
 - Add GXBB specific codepath (untested)

* General:
 - Fix headers license (Thanks chewitt & narmstrong)
 - Only allow one decoding instance at a time