media: meson: add v4l2 m2m video decoder driver
Amlogic SoCs feature a powerful video decoder unit able to
decode many formats, with a performance of usually up to 4k60.

This is a driver for this IP that is based around the v4l2 m2m framework.

It features decoding for:
- MPEG 1/2/4, H.264, MJPEG, (partial) HEVC 8-bit

Even though they are supported in hardware, it doesn't leverage support for:
- HEVC 10-bit, VP9, VC1 (all those are in TODOs)

The code was made in such a way to allow easy inclusion of those formats in the future.
Supported SoCs are: GXBB (S905), GXL (S905X/W/D), GXM (S912)

Specifically, two of the vdec units are enabled within this driver:

- VDEC_1: can decode MPEG 1/2/4, H.264, MJPEG
- VDEC_HEVC: can decode HEVC

There is also a hardware bitstream parser (ESPARSER) that is handled here.